+++ Sources/TuistSupportTesting/TestCase/Fixtures.swift
@@ -1,0 +1,9 @@
+import Path
+
+public enum Fixtures {
+    public static var directory: AbsolutePath {
+        // swiftlint:disable:next force_try
+        return try! AbsolutePath(validating: "\(#file)").parentDirectory.parentDirectory.parentDirectory.parentDirectory
+            .appending(component: "fixtures")
+    }
+}
